Two spellings of one fixture slot throw, reruns cannot go negative, the tie-break is pinned, and the distilled-speed claim matches the table
The case-folded extension match let foo.wav and foo.WAV share a slot with the walk's order picking the winner; the per-stem map now merges with a throw, FixtureError.twoOfAKind, and a test builds a case-sensitive disk image with hdiutil since APFS as shipped cannot hold both names. LatencyHarness takes reruns as UInt so a negative count is unrepresentable rather than trapped. A transposition test pins the documented substitution-first tie-break. README and PROJECT.md no longer say the distilled models are faster than the default: only their turbo variant is, and the plain 594MB one is the slowest of the six.

+++ Sources/lowtalker/BenchCommand.swift
@@ -38,7 +38,7 @@ struct BenchCommand: AsyncParsableCommand {
         for model in models {
             let reporter = PhaseReporter()
             print("model \(model)", to: &stderr)
-            let report = try await LatencyHarness.measure(fixtures, reruns: runs - 1) {
+            let report = try await LatencyHarness.measure(fixtures, reruns: UInt(runs - 1)) {
                 try await WhisperKitTranscriber.load(model, from: store, phase: reporter.report)
             }
             for result in report.fixtures {
